@@ -1,7 +1,20 @@
 package com.kob.backend.service;
 
-import com.baomidou.mybatisplus.extension.service.IService;
-import com.kob.backend.entity.Bot;
 
-public interface IBotService extends IService<Bot> {
+import com.kob.backend.entity.Bot;
+import org.springframework.data.domain.Page;
+
+import java.util.Map;
+
+public interface IBotService{
+
+     Bot save(Bot bot);
+
+     Bot getById(Long id);
+
+     boolean deleteById(Long id);
+
+     boolean updateById(Bot bot);
+
+     Page<Bot> page(Map<String, Object> params);
 }
