@@ -1,13 +1,12 @@
 package com.kob.backend.service;
 
-import com.baomidou.mybatisplus.extension.service.IService;
 import com.kob.backend.dto.UserParam;
 import com.kob.backend.entity.User;
 
 import java.util.List;
 import java.util.Map;
 
-public interface IUserService extends IService<User> {
+public interface IUserService{
 
 
     Map<String, String> login(UserParam userParam) ;
@@ -15,4 +14,6 @@ public interface IUserService extends IService<User> {
     Map<String, String> register(UserParam userParam) ;
 
     User info(Long id) ;
+
+    User getById(Long id) ;
 }
