@@ -1,27 +1,31 @@
 package com.kob.backend.service.serviceImpl;
 
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.kob.backend.common.PageUtils;
-import com.kob.backend.common.Query;
 import com.kob.backend.entity.User;
-import com.kob.backend.mapper.UserMapper;
+import com.kob.backend.repository.UserRepository;
 import com.kob.backend.service.RankListService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import jakarta.annotation.Resource;
 import java.util.Map;
 
 
 @Service
-public class RankListServiceImpl extends ServiceImpl<UserMapper, User> implements RankListService {
+public class RankListServiceImpl implements RankListService {
+
+    @Resource
+    private UserRepository userRepository;
 
     @Override
     public  PageUtils<User> getRankList(Map<String, Object> params) {
-        LambdaQueryWrapper<User> qw = new LambdaQueryWrapper<>();
-        qw.orderByDesc(User::getRating);
-        IPage<User> page = this.page(new Query<User>().getPage(params), qw);
-        return new PageUtils<>(page);
+        int page = (int) params.getOrDefault("page", 1);
+        int limit = (int) params.getOrDefault("limit", 10);
+        PageRequest pageRequest = PageRequest.of(page, limit, Sort.by(Sort.Order.desc("rating")));
+        Page<User> all = userRepository.findAll(pageRequest);
+        return new PageUtils<>(all);
     }
 
 }
