@@ -1,48 +1,48 @@
 <template>
   <content-field>
     <a-form
-      class="form"
-      :model="data"
-      autocomplete="off"
-      @submit.prevent="submit"
-      :label-col="{ span: 6 }"
-      :wrapper-col="{ span: 18 }"
+        class="form"
+        :model="data"
+        autocomplete="off"
+        @submit.prevent="submit"
+        :label-col="{ span: 6 }"
+        :wrapper-col="{ span: 18 }"
     >
       <a-form-item
-        label="用户名"
-        :rules="[{ required: true, message: 'Please input your username!' }]"
+          label="用户名"
+          :rules="[{ required: true, message: 'Please input your username!' }]"
       >
         <a-input
-          v-model="data.username"
-          type="text"
-          class="form-control"
-          id="username"
-          placeholder="请输入用户名"
+            v-model:value="data.username"
+            type="text"
+            class="form-control"
+            id="username"
+            placeholder="请输入用户名"
         />
       </a-form-item>
       <a-form-item
-        label="密码"
-        :rules="[{ required: true, message: 'Please input your username!' }]"
+          label="密码"
+          :rules="[{ required: true, message: 'Please input your username!' }]"
       >
         <a-input
-          v-model="data.password"
-          type="password"
-          class="form-control"
-          id="password"
-          placeholder="请输入密码"
+            v-model:value="data.password"
+            type="password"
+            class="form-control"
+            id="password"
+            placeholder="请输入密码"
         />
       </a-form-item>
 
       <a-form-item
-        label="确认密码"
-        :rules="[{ required: true, message: 'Please input your username!' }]"
+          label="确认密码"
+          :rules="[{ required: true, message: 'Please input your username!' }]"
       >
         <a-input
-          v-model="data.password"
-          type="password"
-          class="form-control"
-          id="confirmpassword"
-          placeholder="请再次输入密码"
+            v-model:value="data.confirmPassword"
+            type="password"
+            class="form-control"
+            id="confirmpassword"
+            placeholder="请再次输入密码"
         />
       </a-form-item>
       <a-button style="width: 100%" type="primary" html-type="submit">
@@ -54,8 +54,9 @@
 
 <script setup>
 import ContentField from "@/components/ContentField.vue";
-import { reactive, ref } from "vue";
-import { register } from "@/api/auth";
+import {reactive, ref} from "vue";
+import {register} from "@/api/auth";
+import router from "@/router/routes.js";
 
 const data = reactive({
   username: "",
@@ -66,12 +67,14 @@ const errorMessage = ref("");
 
 const submit = () => {
   register(data)
-    .then((res) => {
-      console.log(res);
-    })
-    .catch((e) => {
-      errorMessage.value = e.response.data.message;
-    });
+      .then((res) => {
+        console.log(res);
+        router.push({name: "Login"});
+      })
+      .catch((e) => {
+        console.log(e.response);
+        errorMessage.value = e.response.data.message;
+      });
 };
 </script>
 
